move test to cupy test file
